Use VCR for tests.
Using VCR stores the HTTP requests made.  It is configured to only be
active when explicitly used with
`VCR.use_cassette(cassette_name,&block)`.  This configuration is
necessary for the feature specs, as these make requests to localhost
that should not be recorded.  Since we don't know (and don't want to
know) how Travis is running our tests, just disabling VCR for all
requests to localhost was not an option.